Update graph to take audio transcription as input

+++ backend/src/core/tools/calculator.py
@@ -1,12 +1,12 @@
 class CalculatorService:
-    def add(a: float, b: float):
+    def add(a: int, b: int):
         """Add two numbers."""
         return a + b
 
-    def multiply(a: float, b: float):
+    def multiply(a: int, b: int):
         """Multiply two numbers."""
         return a * b
 
-    def divide(a: float, b: float):
+    def divide(a: int, b: int):
         """Divide two numbers."""
         return a / b
